Kubernetes Bàsic/fase 3/product-service/requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
redis==5.0.8
